requirements.txt
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.0
typing-extensions>=4.5.0
dataclasses>=0.6; python_version < '3.7'